Theoretical bloc price calculus

+++ src/main/java/com/tiarintsoa/foam/dto/StockProduitDTO.java
@@ -10,7 +10,7 @@ public class StockProduitDTO extends StockProduitPriceAverageDTO {
     private String nomOrigine;
     private String nomOriginel;
 
-    public StockProduitDTO(String nomProduit, int quantite, double coutProductionUnitaire, double prixVenteUnitaire, String nomOrigine, String nomOriginel) {
+    public StockProduitDTO(String nomProduit, double quantite, double coutProductionUnitaire, double prixVenteUnitaire, String nomOrigine, String nomOriginel) {
         this.nomProduit = nomProduit;
         this.quantite = quantite;
         this.coutProductionUnitaire = coutProductionUnitaire;
