+++ src/main/java/com/tiarintsoa/foam/dto/StockProduitPriceAverageDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 @Data
 public class StockProduitPriceAverageDTO {
     protected String nomProduit;
-    protected int quantite;
+    protected double quantite;
     protected double coutProductionUnitaire;
     protected double prixVenteUnitaire;
 
     public StockProduitPriceAverageDTO() {}
 
-    public StockProduitPriceAverageDTO(String nomProduit, int quantite, double coutProductionUnitaire, double prixVenteUnitaire) {
+    public StockProduitPriceAverageDTO(String nomProduit, double quantite, double coutProductionUnitaire, double prixVenteUnitaire) {
         this.nomProduit = nomProduit;
         this.quantite = quantite;
         this.coutProductionUnitaire = coutProductionUnitaire;
